chore: removing adding method on edgeless block
